Fix Denunceation typo, reorg compatibility code
