+++ core/src/com/unciv/ui/trade/DiplomacyScreen.kt
@@ -16,7 +16,6 @@ import com.unciv.logic.civilization.diplomacy.RelationshipLevel
 import com.unciv.logic.trade.TradeLogic
 import com.unciv.logic.trade.TradeOffer
 import com.unciv.logic.trade.TradeType
-import com.unciv.models.ruleset.Era
 import com.unciv.models.ruleset.ModOptionsConstants
 import com.unciv.models.ruleset.Quest
 import com.unciv.models.ruleset.tile.ResourceType
@@ -582,7 +581,7 @@ class DiplomacyScreen(val viewingCiv:CivilizationInfo):CameraStageBaseScreen() {
                 diplomacyTable.add(researchAgreementButton).row()
             }
 
-            if (!diplomacyManager.hasFlag(DiplomacyFlags.Denunceation)
+            if (!diplomacyManager.hasFlag(DiplomacyFlags.Denunciation)
                 && !diplomacyManager.hasFlag(DiplomacyFlags.DeclarationOfFriendship)
             ) {
                 val denounceButton = "Denounce ([30] turns)".toTextButton()
